requirements.txt
Django==1.7.1
argparse==1.2.1
dj-database-url==0.3.0
waitress==0.8.9
whitenoise==1.0.6
wsgiref==0.1.2
psycopg2==2.5.4
Fabric==1.8.3
django-braces==1.4.0
django-tinymce==1.5.2
